Remove CODE_OF_CONDUCT.md
I am against these kinds of documents because they contain phrases like 'harmful language' that are open to
interpretation. This means the document can be used as a shield or excuse for someone with arbitrary powers
to oppress others by limiting their speech with the threat of punishment.

I'm also against them because they presuppose that we are not adults who can deal with each other in the
real world, and that we need a rule book in order to be pleasant to each other.

Given the insulting nature of these documents and how easily they can be abused, I conclude that it is these
documents themselves that contain the harmful language. Therefore they must be expunged.

Also they tend to cause drama.

I wrote to the team 12+ hours ago: "Unless I receive a compelling reason not to do so within 12 hours,
I will unilaterally remove this document from the repo." - I didn't receive a compelling reason.